added lambert conformal conic

+++ GCTP_CPP/projection.cpp
@@ -11,7 +11,7 @@ m_forInitNeeded(true), m_invInitNeeded(true)
 		m_gctpParams[i] = 0.0;
 }
 
-Projection::Projection ( double gctpParameters[], int units, int datum, int spheroid ): 
+Projection::Projection ( double gctpParameters[], int units, long datum, long spheroid ): 
 m_number(-1), m_unitCode(units), m_datum(datum), m_spheroid(spheroid),
 m_longitude(0.0), m_latitude(0.0), m_x_coord(0.0), m_y_coord(0.0), m_falseEasting(0.0),
 m_falseNorthing(0.0), m_rMajor(0.0), m_rMinor(0.0), m_radius(0.0),
