+++ GCTP_CPP/sinusoidal.h
@@ -17,7 +17,7 @@ class Sinusoidal : public Projection
   
   public:
    Sinusoidal();
-   Sinusoidal(double gctpParameters[], int units = 0, int datum = 0, int spheroid = 0);
+   Sinusoidal(double gctpParameters[],int units = 0, long datum = 0, long spheroid = 0);
    void forward( double lon, double lat, double* x = NULL, double* y = NULL );
    void inverse( double x, double y, double* lon = NULL, double* lat = NULL );
